+++ src/main/java/cybereats/fpmislata/com/tiendaback/domain/service/impl/CategoryPCServiceImpl.java
@@ -4,6 +4,8 @@ import cybereats.fpmislata.com.tiendaback.domain.model.Page;
 import cybereats.fpmislata.com.tiendaback.domain.repository.CategoryPCRepository;
 import cybereats.fpmislata.com.tiendaback.domain.service.CategoryPCService;
 import cybereats.fpmislata.com.tiendaback.domain.service.dto.CategoryPCDto;
+import cybereats.fpmislata.com.tiendaback.exception.BusinessException;
+import cybereats.fpmislata.com.tiendaback.exception.ResourceNotFoundException;
 
 import java.util.Optional;
 
@@ -15,40 +17,52 @@ public class CategoryPCServiceImpl implements CategoryPCService {
     }
 
     @Override
-    public Page<CategoryPCDto> getAll(int page, int size) {
+    public Page<CategoryPCDto> findAll(int page, int size) {
         Page<CategoryPCDto> categoryPCDtoPage = categoryPCRepository.findAll(page, size);
 
         return new Page<>(
                 categoryPCDtoPage.data(),
                 categoryPCDtoPage.pageNumber(),
                 categoryPCDtoPage.pageSize(),
-                categoryPCDtoPage.totalElements()
-        );
+                categoryPCDtoPage.totalElements());
     }
 
     @Override
     public CategoryPCDto getById(Long id) {
         return categoryPCRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("CategoryPC not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("CategoryPC not found"));
     }
 
     @Override
     public Optional<CategoryPCDto> findById(Long id) {
-        return Optional.ofNullable(categoryPCRepository.findById(id).orElseThrow(() -> new RuntimeException("Booking not found")));
+        return Optional.ofNullable(categoryPCRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Booking not found")));
     }
 
     @Override
     public CategoryPCDto create(CategoryPCDto categoryPCDto) {
+        Optional<CategoryPCDto> categoryPCDtoOptional = categoryPCRepository.findById(categoryPCDto.id());
+        if (categoryPCDtoOptional.isPresent()) {
+            throw new BusinessException("CategoryPC already exists");
+        }
         return categoryPCRepository.save(categoryPCDto);
     }
 
     @Override
     public CategoryPCDto update(CategoryPCDto categoryPCDto) {
+        Optional<CategoryPCDto> categoryPCDtoOptional = categoryPCRepository.findById(categoryPCDto.id());
+        if (categoryPCDtoOptional.isEmpty()) {
+            throw new ResourceNotFoundException("CategoryPC not found");
+        }
         return categoryPCRepository.save(categoryPCDto);
     }
 
     @Override
     public void deleteById(Long id) {
+        Optional<CategoryPCDto> categoryPCDtoOptional = categoryPCRepository.findById(id);
+        if (categoryPCDtoOptional.isEmpty()) {
+            throw new ResourceNotFoundException("CategoryPC not found");
+        }
         categoryPCRepository.deleteById(id);
     }
 }
